requirements.txt
PyGObject>=3.42.0
PyYAML>=6.0
